+++ screens/org-onboarding-screens/FileOrderDocumentation.md
@@ -1,6 +1,6 @@
 # Order of Organization-Onboarding screens
 
-Dear Key Conservation app programmers, 
+Dear Key Conservation app programmers,
 
 Because there are so many onboarding screens, we have provided you with this list showing the order that these screens go in:
 
@@ -16,12 +16,12 @@ Because there are so many onboarding screens, we have provided you with this lis
 10. VerifyDocumentationScreen
 11. ReviewYourInfoScreen
 12. ToExpectNextCreateProfile
-13. OrganizationSurveryScreen
+13. AccountScreen
 
 **These are completed and hooked up but not currently in use**
+
 11. AlmostDoneScreen
 12. ThankYouScreen
 
-
-Yours Truly, 
+Yours Truly,
 Julie Christine Gumerman :D
